Better manage the event loop group and fix stopping the attack
It makes more sense to have one central event loop, so we don't need to worry about too many being created. We now init the eventloop with the number of bots, so we have enough threads premade and ready.

Further, the reconnect and start logic now uses the event loop group to be able to shut down together with the group centrally.