Cleanup attack starting and also make max concurrent connects configurable
Concurrent connects mean the amount of bots that can try to connect at the same time. This feature reflects the join delay as well.
If join delay for all bots is 2 seconds and one bot takes 3 seconds to join then the bot after will wait for the first bot to finish connecting and so everything gets delayed. This setting allows configuring how many bots can connect at once. If two is set then for example the second bot could try to connect even if the first hasn't finished joining yet.