Implement inventory projection and also fixup level state branching
If we really used projected level state like now, the lookup cache would influence the hash, which would cause two states to be different just because one had different lookups.
This would HEAVILY impact performance. In the future though to improve lookup speeds if slow, cache may be reimplemented, but so that it doesn't affect the states hash.